+++ benchmark/test_logical_not.py
@@ -10,3 +10,13 @@ def test_logical_not():
         op_name="logical_not", torch_op=torch.logical_not, dtypes=consts.FLOAT_DTYPES
     )
     bench.run()
+
+
+@pytest.mark.logical_not_
+def test_logical_not_():
+    bench = base.UnaryPointwiseBenchmark(
+        op_name="logical_not_",
+        torch_op=lambda a: a.logical_not_(),
+        dtypes=consts.INT_DTYPES + consts.BOOL_DTYPES,
+    )
+    bench.run()
